move: UI model presentation으로 이동
