+++ presentation/src/main/res/drawable/ic_arrow_left.xml
