+++ presentation/src/main/res/drawable/ic_arrow_right.xml
