+++ presentation/src/main/res/drawable/ic_arrow_up.xml
