+++ presentation/src/main/res/drawable/ic_book.xml
