+++ presentation/src/main/res/drawable/ic_category_default.xml
