+++ presentation/src/main/res/drawable/ic_chevron_right.xml
