+++ presentation/src/main/res/drawable/ic_clock.xml
