+++ presentation/src/main/res/drawable/ic_close.xml
