+++ presentation/src/main/res/drawable/ic_focus.xml
