+++ presentation/src/main/res/drawable/ic_memo.xml
