+++ presentation/src/main/res/drawable/ic_menu.xml
