+++ presentation/src/main/res/drawable/ic_minus.xml
