+++ presentation/src/main/res/drawable/ic_monitor.xml
