+++ presentation/src/main/res/drawable/ic_pen.xml
