+++ presentation/src/main/res/drawable/ic_play.xml
