+++ presentation/src/main/res/drawable/ic_plus.xml
