+++ presentation/src/main/res/drawable/ic_rest.xml
